change Employee.ReviewScore back to byte

+++ Model/Employee.cs
@@ -5,7 +5,7 @@ namespace Model
     public class Employee : User
     {
         public Category Expertise { get; set; }
-        public int ReviewScore { get; set; }
+        public byte ReviewScore { get; set; }
         public int TicketsSolved { get; set; }
 
         public Employee(string firstName, string lastName, DateTime birthDate, string email, string password, Category expertise)
